third_party/blink/web_tests/platform/win/external/wpt/url/url-constructor.any.worker_include=file-expected.txt

This is a testharness.js-based test.
Found 126 tests; 74 PASS, 52 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Loading dataâ€¦
FAIL Parsing: <file:/example.com/> against <http://example.org/foo/bar> assert_equals: href expected "file:///example.com/" but got "file://example.com/"
PASS Parsing: <file://example:1/> without base
PASS Parsing: <file://example:test/> without base
PASS Parsing: <file://example%/> without base
PASS Parsing: <file://[example]/> without base
FAIL Parsing: <file:c:\foo\bar.html> against <file:///tmp/mock/path> assert_equals: href expected "file:///c:/foo/bar.html" but got "file:///C:/foo/bar.html"
FAIL Parsing: <  File:c|////foo\bar.html> against <file:///tmp/mock/path> assert_equals: href expected "file:///c:////foo/bar.html" but got "file:///C:////foo/bar.html"
PASS Parsing: <C|/foo/bar> against <file:///tmp/mock/path>
PASS Parsing: </C|\foo\bar> against <file:///tmp/mock/path>
PASS Parsing: <//C|/foo/bar> against <file:///tmp/mock/path>
PASS Parsing: <//server/file> against <file:///tmp/mock/path>
PASS Parsing: <\\server\file> against <file:///tmp/mock/path>
PASS Parsing: </\server/file> against <file:///tmp/mock/path>
PASS Parsing: <file:///foo/bar.txt> against <file:///tmp/mock/path>
PASS Parsing: <file:///home/me> against <file:///tmp/mock/path>
PASS Parsing: <//> against <file:///tmp/mock/path>
PASS Parsing: <///> against <file:///tmp/mock/path>
PASS Parsing: <///test> against <file:///tmp/mock/path>
PASS Parsing: <file://test> against <file:///tmp/mock/path>
FAIL Parsing: <file://localhost> against <file:///tmp/mock/path> assert_equals: href expected "file:///" but got "file://localhost/"
FAIL Parsing: <file://localhost/> against <file:///tmp/mock/path> assert_equals: href expected "file:///" but got "file://localhost/"
FAIL Parsing: <file://localhost/test> against <file:///tmp/mock/path> assert_equals: href expected "file:///test" but got "file://localhost/test"
PASS Parsing: <test> against <file:///tmp/mock/path>
PASS Parsing: <file:test> against <file:///tmp/mock/path>
FAIL Parsing: <file:/example.com/> without base assert_equals: href expected "file:///example.com/" but got "file://example.com/"
FAIL Parsing: <file:...> against <http://www.example.com/test> assert_equals: href expected "file:///..." but got "file://.../"
FAIL Parsing: <file:..> against <http://www.example.com/test> assert_equals: href expected "file:///" but got "file://../"
FAIL Parsing: <file:a> against <http://www.example.com/test> assert_equals: href expected "file:///a" but got "file://a/"
PASS Parsing: <file:///C%3A/> without base
PASS Parsing: <file:///C%7C/> without base
PASS Parsing: <file://%43%3A> without base
PASS Parsing: <file://%43%7C> without base
PASS Parsing: <file://%43|> without base
PASS Parsing: <file://C%7C> without base
PASS Parsing: <file://%43%7C/> without base
PASS Parsing: <pix/submit.gif> against <file:///C:/Users/Domenic/Dropbox/GitHub/tmpvar/jsdom/test/level2/html/files/anchor.html>
PASS Parsing: <..> against <file:///C:/>
PASS Parsing: <..> against <file:///>
PASS Parsing: </> against <file:///C:/a/b>
PASS Parsing: </> against <file://h/C:/a/b>
PASS Parsing: </> against <file://h/a/b>
FAIL Parsing: <//d:> against <file:///C:/a/b> assert_equals: href expected "file:///d:" but got "file:///D:"
FAIL Parsing: <//d:/..> against <file:///C:/a/b> assert_equals: href expected "file:///d:/" but got "file:///D:/"
PASS Parsing: <..> against <file:///ab:/>
PASS Parsing: <..> against <file:///1:/>
PASS Parsing: <> against <file:///test?test#test>
PASS Parsing: <file:> against <file:///test?test#test>
PASS Parsing: <?x> against <file:///test?test#test>
PASS Parsing: <file:?x> against <file:///test?test#test>
PASS Parsing: <#x> against <file:///test?test#test>
PASS Parsing: <file:#x> against <file:///test?test#test>
FAIL Parsing: <file:\\//> without base assert_equals: href expected "file:////" but got "file:///"
FAIL Parsing: <file:\\\\> without base assert_equals: href expected "file:////" but got "file:///"
FAIL Parsing: <file:\\\\?fox> without base Failed to construct 'URL': Invalid URL
FAIL Parsing: <file:\\\\#guppy> without base Failed to construct 'URL': Invalid URL
PASS Parsing: <file://spider///> without base
FAIL Parsing: <file:\\localhost//> without base assert_equals: href expected "file:////" but got "file://localhost//"
PASS Parsing: <file:///localhost//cat> without base
FAIL Parsing: <file://\/localhost//cat> without base assert_equals: href expected "file:////localhost//cat" but got "file://localhost//cat"
FAIL Parsing: <file://localhost//a//../..//> without base assert_equals: href expected "file://///" but got "file://localhost///"
FAIL Parsing: </////mouse> against <file:///elephant> assert_equals: href expected "file://///mouse" but got "file://mouse/"
PASS Parsing: <\//pig> against <file://lion/>
FAIL Parsing: <\/localhost//pig> against <file://lion/> assert_equals: href expected "file:////pig" but got "file://localhost//pig"
FAIL Parsing: <//localhost//pig> against <file://lion/> assert_equals: href expected "file:////pig" but got "file://localhost//pig"
PASS Parsing: </..//localhost//pig> against <file://lion/>
PASS Parsing: <file://> against <file://ape/>
PASS Parsing: </rooibos> against <file://tea/>
PASS Parsing: </?chai> against <file://tea/>
FAIL Parsing: <C|> against <file://host/dir/file> assert_equals: href expected "file://host/C:" but got "file:///C:"
FAIL Parsing: <C|> against <file://host/D:/dir1/dir2/file> assert_equals: href expected "file://host/C:" but got "file:///C:"
FAIL Parsing: <C|#> against <file://host/dir/file> assert_equals: href expected "file://host/C:#" but got "file:///C:#"
FAIL Parsing: <C|?> against <file://host/dir/file> assert_equals: href expected "file://host/C:?" but got "file:///C:?"
FAIL Parsing: <C|/> against <file://host/dir/file> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
FAIL Parsing: <C|
/> against <file://host/dir/file> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
FAIL Parsing: <C|\> against <file://host/dir/file> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
PASS Parsing: <C> against <file://host/dir/file>
FAIL Parsing: <C|a> against <file://host/dir/file> assert_equals: href expected "file://host/dir/C|a" but got "file:///C:/a"
FAIL Parsing: </c:/foo/bar> against <file:///c:/baz/qux> assert_equals: href expected "file:///c:/foo/bar" but got "file:///C:/foo/bar"
FAIL Parsing: </c|/foo/bar> against <file:///c:/baz/qux> assert_equals: href expected "file:///c:/foo/bar" but got "file:///C:/foo/bar"
FAIL Parsing: <file:\c:\foo\bar> against <file:///c:/baz/qux> assert_equals: href expected "file:///c:/foo/bar" but got "file:///C:/foo/bar"
FAIL Parsing: </c:/foo/bar> against <file://host/path> assert_equals: href expected "file://host/c:/foo/bar" but got "file:///C:/foo/bar"
PASS Parsing: <file://example.net/C:/> without base
PASS Parsing: <file://1.2.3.4/C:/> without base
PASS Parsing: <file://[1::8]/C:/> without base
FAIL Parsing: <C|/> against <file://host/> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
FAIL Parsing: </C:/> against <file://host/> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
FAIL Parsing: <file:C:/> against <file://host/> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
FAIL Parsing: <file:/C:/> against <file://host/> assert_equals: href expected "file://host/C:/" but got "file:///C:/"
PASS Parsing: <//C:/> against <file://host/>
PASS Parsing: <file://C:/> against <file://host/>
PASS Parsing: <///C:/> against <file://host/>
PASS Parsing: <file:///C:/> against <file://host/>
PASS Parsing: <file:/C|/> without base
PASS Parsing: <file://C|/> without base
PASS Parsing: <file:> without base
FAIL Parsing: <file:?q=v> without base Failed to construct 'URL': Invalid URL
FAIL Parsing: <file:#frag> without base Failed to construct 'URL': Invalid URL
PASS Parsing: <file:///Y:> without base
PASS Parsing: <file:///Y:/> without base
PASS Parsing: <file:///./Y> without base
PASS Parsing: <file:///./Y:> without base
FAIL Parsing: <file:///y:> without base assert_equals: href expected "file:///y:" but got "file:///Y:"
FAIL Parsing: <file:///y:/> without base assert_equals: href expected "file:///y:/" but got "file:///Y:/"
PASS Parsing: <file:///./y> without base
FAIL Parsing: <file:///./y:> without base assert_equals: href expected "file:///y:" but got "file:///Y:"
FAIL Parsing: <file://localhost//a//../..//foo> without base assert_equals: href expected "file://///foo" but got "file://localhost///foo"
FAIL Parsing: <file://localhost////foo> without base assert_equals: href expected "file://////foo" but got "file://localhost////foo"
FAIL Parsing: <file:////foo> without base assert_equals: href expected "file:////foo" but got "file://foo/"
PASS Parsing: <file:///one/two> against <file:///>
FAIL Parsing: <file:////one/two> against <file:///> assert_equals: href expected "file:////one/two" but got "file://one/two"
PASS Parsing: <//one/two> against <file:///>
PASS Parsing: <///one/two> against <file:///>
FAIL Parsing: <////one/two> against <file:///> assert_equals: href expected "file:////one/two" but got "file://one/two"
PASS Parsing: <file:///.//> against <file:////>
FAIL Parsing: <file:.//p> without base assert_equals: href expected "file:////p" but got "file://.//p"
FAIL Parsing: <file:/.//p> without base assert_equals: href expected "file:////p" but got "file://.//p"
PASS Parsing: <10.0.0.7:8080/foo.html> against <file:///some/dir/bar.html>
PASS Parsing: <a!@$*=/foo.html> against <file:///some/dir/bar.html>
PASS Parsing: <file://aÂ­b/p> without base
PASS Parsing: <file://a%C2%ADb/p> without base
FAIL Parsing: <file://loCð€ð‹ð‡ð¨ð¬ð­/usr/bin> without base assert_equals: href expected "file:///usr/bin" but got "file://localhost/usr/bin"
PASS Parsing: <file://Â­/p> without base
PASS Parsing: <file://%C2%AD/p> without base
FAIL Parsing: <file://xn--/p> without base assert_throws_js: function "function() {
          new URL(expected.input, base);
        }" did not throw
PASS Parsing: <> without base
Harness: the test ran to completion.
